// src/app/model/jobs.ts
export class Jobs {
   id: number;
   areaID: number;
   jobsTitle: string;
   jobsStatus: string;
   company: string;
   thumbnail: string;
   jobsLocation: string;
   jobsMail: string;
   jobsDescription: string;
   jobsResponsibilities: string;
   salary: string;
   jobsType: string;
   jobsSkill: string;
   createdDate: Date;
   modifiedDate: Date;
   createdBy: Date;
   modifiedBy: Date;
   status: number;
}